Implement the side panel
The side panel is displayed on all pages of the Dash app and contains,
among other things, links to the other pages.

So far the side panel contains the correct icons and links which can be
activated. What remains is implementing a button for toggling the side
panel visibility and adding correct behavior for the links