Create first mockup of dashboard
Created the layout of the dashboard. Also made it so that the state of
the side panel is stored and shared to the dashboard for smooth resizing
when toggling the side panel.

The behavior is not good on mobile though where the side panel should
always be placed on top of the main content and thus no padding added to
the dashboard.