Implement top row of Reach component
Top row of Reach component now diplays dummy data and also the time span
toggler.

Class for activating a specific time span contains attributes which
invoke the !important modifier which should be avoided. This can be
changed in the future. Also when activating a time span the font doesn't
change to bold because it causes the text box to change size and cause
an uggly jump in the UI. This too can be fixed in the future.